+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "ku-progress-bar",
-  "version": "1.0.0",
+  "version": "0.1.0",
   "description": "",
   "main": "dist/index.js",
   "scripts": {
