+++ src/lib/bar.ts
@@ -1,126 +1,66 @@
-import { IBarOptions } from './interfaces/bar-options.interface';
+import { TerminalTty } from './terminals/terminal-tty';
+import { BarItem } from './bar-item';
 import { Progress } from './progress';
 
-export interface IFormatters {
-  // FIXME: update index to progress
-  [key: string]: (str: string, index: number, progresses: Progress[]) => string;
-}
-
-export interface IDataProviders {
-  [key: string]: (progress: Progress, progresses: Progress[]) => string;
-}
-
-export interface IParams {
-  template?: string;
-  options?: Partial<IBarOptions>;
-  formatters?: IFormatters;
-  dataProviders?: IDataProviders;
-}
-
-// TODO: handle soft close during process.exit(0);
-
 export class Bar {
+  protected items: BarItem[] = [];
+  protected isStarted = false;
+  protected nextUpdate: null | Promise<never> = null;
 
-  protected template: string = `[{bar}] {percentage} ETA: {eta} speed: {speed} duration: {duration} {value}/{total}`;
-  protected options: IBarOptions = {
-    completeChar: '=',
-    resumeChar: '-',
-    width: 40,
-    glue: '',
-  };
-  protected formatters!: IFormatters;
-  protected dataProviders!: IDataProviders;
+  public constructor(
+    protected terminal = new TerminalTty(),
+    ) { // FIXME: terminal should be interface
+  }
 
-  public constructor(protected progresses: Progress[], params?: IParams) {
-    this.template = params?.template ?? this.template;
-    this.options = { ...this.options, ...params?.options };
-    this.formatters = params?.formatters ?? {};
-    this.dataProviders = {
-      bars: (progress, progresses) => this.renderBars(progresses),
-      bar: (progress) => this.bar(progress.getProgress()),
-      speed: (progress) => Math.round(progress.getEta().getSpeed()) + '/s',
-      eta: (progress) => progress.getEta().getEtaS() + 's',
-      value: (progress) => progress.getValue().toString(),
-      total: (progress) => progress.getTotal().toString(),
-      percentage: (progress) => Math.round(progress.getProgress() * 100) + '%',
-      duration: (progress) => Math.round(progress.getEta().getDurationMs() / 1000) + 's',
-      ...params?.dataProviders,
+  public add(bar: BarItem) {
+    this.items.push(bar)
+    if (this.isStarted) {
+      this.addListenerToProgress(bar)
     }
+    return this;
   }
 
-  public getProgresses() {
-    return this.progresses;
+  public removeByProgress(progress: Progress) {
+    this.items = this.items.filter(
+      item => !item.getProgresses().find(p => p == progress)
+    )
+    this.refresh();
   }
 
-  public renderBars(progresses: Progress[]): string {
-    const { resumeChar, width, glue } = this.options;
-    const lines = [];
-
-    const leftLength = progresses
-      .map((item, index) => ({
-        size: Math.round(item.getProgress() * width),
-        item,
-        index,
-      }))
-      .sort((a, b) => Math.sign(a.size - b.size))
-      .reduce((prev, current) => {
-        const length = current.size - prev.size;
-        if (length > 0) {
-          let line = this.getBarParts(length > width ? width : length).done;
-          const item = current.item;
-          const formatter = this.formatters[`${item.getTag()}_bar`] ?? this.formatters['bar'];
-          if (formatter) {
-            line = formatter(line, current.index, progresses);
-          }
-          lines.push(line);
-        }
-        return current;
-      }, { size: 0 });
-
-    if (width - leftLength.size > 0) {
-      lines.push(resumeChar.repeat(width - leftLength.size));
-    }
-    return lines.join(glue);
+  public render() {
+    const lines = this.items.map( bar => {
+      return bar.render();
+    })
+    this.terminal.write(lines.join('\n') + '\n');
   }
 
-  public render(): string {
-    return this.template.replace(/{([^}]+)}/g, (match, prop) => {
-      const [property, tag] = prop.split('_').reverse();
-      const index = tag ? this.progresses.findIndex(p => p.getTag() === tag) : 0;
-      if (index < 0) return match;
-      const progress = this.progresses[index];
-      const value = this.getDataValue(property, progress);
-      if (value === null) {
-        return match;
-      }
-      if (this.formatters[prop]) {
-        return this.formatters[prop](value, index, this.progresses);
-      }
-      return value;
+  public log(logFunction: () => void) {
+    this.terminal.clear();
+    logFunction();
+    this.terminal.refresh();
+  }
+
+  public start() {
+    this.isStarted = true;
+    this.items.forEach(item => this.addListenerToProgress(item))
+    this.render();
+  }
+
+  protected addListenerToProgress(item: BarItem) {
+    item.getProgresses().forEach(progress => {
+      progress.on('update', this.refresh);
     });
   }
 
-  public getBarParts(size: number): { left: string; done: string } {
-    return {
-      done: this.options.completeChar.repeat(size),
-      left: this.options.resumeChar.repeat(this.options.width - size)
+  protected refresh = () => {
+    if (!this.isStarted || this.nextUpdate) {
+      return;
     }
+    this.nextUpdate = new Promise(resolve => {
+      setTimeout(() => {
+        this.nextUpdate = null;
+        this.render();
+      }, 50); // FIXME: config or constant
+    })
   }
-
-  public bar(progress: number): string {
-    const size =  Math.round(progress * this.options.width);
-    const parts = this.getBarParts(size);
-    return `${parts.done}${this.options.glue}${parts.left}`;
-  }
-
-  protected getDataValue = (key: string, item: Progress): string | null => {
-    // const map: { [key: string]: () => string } = {
-    // }
-    const payload = item.getPayload();
-    let value = payload[key] ?? null;
-    value = (value === null && this.dataProviders[key]) ? this.dataProviders[key](item, this.progresses) : value;
-    if (value === null) return value;
-    return value;
-  };
-
 }
